+++ idf_component_manager/core_utils.py
@@ -1,6 +1,6 @@
 # SPDX-FileCopyrightText: 2022 Espressif Systems (Shanghai) CO LTD
 # SPDX-License-Identifier: Apache-2.0
-
+import os
 import re
 from pathlib import Path
 
@@ -8,6 +8,7 @@ from tqdm import tqdm
 
 from idf_component_tools.constants import DEFAULT_NAMESPACE
 from idf_component_tools.errors import ComponentModifiedError, FatalError
+from idf_component_tools.file_tools import copy_directories, filtered_paths
 from idf_component_tools.hash_tools import HASH_FILENAME
 from idf_component_tools.manifest import Manifest
 from idf_component_tools.semver import SimpleSpec
@@ -83,3 +84,75 @@ def parse_example(example):  # type: (str) -> tuple[str, str, str]
             'Invalid version specification: "{}". Please use format like ">=1" or "*".'.format(version_spec))
 
     return '{}/{}'.format(namespace, component), version_spec, example_name
+
+
+def collect_directories(dir_path):  # type: (Path) -> list[str]
+    directories = []  # type: list[str]
+    if not dir_path.is_dir():
+        return directories
+
+    for directory in os.listdir(str(dir_path)):
+        if directory.startswith('.') or not (dir_path / directory).is_dir():
+            continue
+
+        directories.append(directory)
+
+    return directories
+
+
+def detect_duplicate_examples(example_folders, example_path, example_name):  # type ()
+    for key, value in example_folders.items():
+        if example_name in value:
+            return key, example_path, example_name
+    return
+
+
+def copy_examples_folders(
+        examples_manifest,  # type: list[dict[str,str]]
+        working_path,  # type: Path
+        dist_dir,  # type: Path
+        include=None,  # type: set[str] | None
+        exclude=None,  # type: set[str] | None
+):  # type: (...) -> None
+    examples_path = working_path / 'examples'
+    example_folders = {'examples': collect_directories(examples_path)}
+    error_paths = []
+    duplicate_paths = []
+    for example_info in examples_manifest:
+        example_path = example_info['path']
+        example_name = Path(example_path).name
+        full_example_path = working_path / example_path
+
+        if not full_example_path.is_dir():
+            error_paths.append(str(full_example_path))
+            continue
+
+        if example_path in example_folders.keys():
+            raise FatalError(
+                'Some paths in the `examples` block in the manifest are listed multiple times: {}. '
+                'Please make paths unique and delete duplicate paths'.format(example_path))
+
+        duplicates = detect_duplicate_examples(example_folders, example_path, example_name)
+        if duplicates:
+            duplicate_paths.append(duplicates)
+            continue
+
+        example_folders[example_path] = [example_name]
+
+        paths = filtered_paths(full_example_path, include=include, exclude=exclude)
+        copy_directories(str(full_example_path), str(dist_dir / 'examples' / example_name), paths)
+
+    if error_paths:
+        raise FatalError(
+            "Example directory doesn't exist: {}.\n"
+            'Please check the path of the custom example folder in `examples` field in `idf_component.yml` file'.format(
+                ', '.join(error_paths)))
+
+    if duplicate_paths:
+        error_messages = []
+        for first_path, second_path, example_name in duplicate_paths:
+            error_messages.append(
+                'Examples from "{}" and "{}" have the same name: {}.'.format(first_path, second_path, example_name))
+        error_messages.append('Please rename one of them, or delete if there are the same')
+
+        raise FatalError('\n'.join(error_messages))