+++ tests/fixtures/components/cmp_for_examples/cmp.c
@@ -1,0 +1,11 @@
+/*
+ * SPDX-FileCopyrightText: 2022 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+#include <stdio.h>
+
+void cmp_hello()
+{
+        printf("Hello from CMP version 1.0.0\n");
+}