+++ tests/fixtures/components/cmp_for_examples/custom_example_path/cmp_ex/main/cmp_ex.c
@@ -1,0 +1,11 @@
+/*
+ * SPDX-FileCopyrightText: 2022 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+#include "cmp.h"
+
+void app_main(void)
+{
+    cmp_hello();
+}